Revert "Add Fault-Tolerance tests & generalise the Multi-DP integration tests"
