+++ clib/clib_mininet_test.py
@@ -14,4 +14,4 @@ from clib_mininet_test_main import test_main
 import clib_mininet_tests
 
 if __name__ == '__main__':
-    test_main([clib_mininet_tests.__name__])
+    test_main(clib_mininet_tests.__name__)
