+++ clib/mininet_test_base.py
@@ -157,8 +157,6 @@ class FaucetTestBase(unittest.TestCase):
 
     def first_switch(self):
         """Return first switch by name order."""
-        if not self.switches_name_ordered():
-            return None
         return self.switches_name_ordered()[0]
 
     def rand_dpid(self):
@@ -389,20 +387,6 @@ class FaucetTestBase(unittest.TestCase):
     def hostns(self, host):
         return '%s' % host.name
 
-    def dump_switch_flows(self, switch):
-        """ """
-        for dump_cmd in (
-                'dump-flows', 'dump-groups', 'dump-meters',
-                'dump-group-stats', 'dump-ports', 'dump-ports-desc',
-                'meter-stats'):
-            switch_dump_name = os.path.join(self.tmpdir, '%s-%s.log' % (switch.name, dump_cmd))
-            # TODO: occasionally fails with socket error.
-            switch.cmd('%s %s %s > %s' % (self.OFCTL, dump_cmd, switch.name, switch_dump_name),
-                       success=None)
-        for other_cmd in ('show', 'list controller', 'list manager'):
-            other_dump_name = os.path.join(self.tmpdir, '%s.log' % other_cmd.replace(' ', ''))
-            switch.cmd('%s %s > %s' % (self.VSCTL, other_cmd, other_dump_name))
-
     def tearDown(self, ignore_oferrors=False):
         """Clean up after a test.
            ignore_oferrors: return OF errors rather than failing"""
@@ -410,13 +394,21 @@ class FaucetTestBase(unittest.TestCase):
             for host in self.hosts_name_ordered()[:1]:
                 if self.get_host_netns(host):
                     self.quiet_commands(host, ['ip netns del %s' % self.hostns(host)])
-        first_switch = self.first_switch()
-        if first_switch:
-            self.first_switch().cmd('ip link > %s' % os.path.join(self.tmpdir, 'ip-links.log'))
+        self.first_switch().cmd('ip link > %s' % os.path.join(self.tmpdir, 'ip-links.log'))
         switch_names = []
         for switch in self.net.switches:
             switch_names.append(switch.name)
-            self.dump_switch_flows(switch)
+            for dump_cmd in (
+                    'dump-flows', 'dump-groups', 'dump-meters',
+                    'dump-group-stats', 'dump-ports', 'dump-ports-desc',
+                    'meter-stats'):
+                switch_dump_name = os.path.join(self.tmpdir, '%s-%s.log' % (switch.name, dump_cmd))
+                # TODO: occasionally fails with socket error.
+                switch.cmd('%s %s %s > %s' % (self.OFCTL, dump_cmd, switch.name, switch_dump_name),
+                           success=None)
+            for other_cmd in ('show', 'list controller', 'list manager'):
+                other_dump_name = os.path.join(self.tmpdir, '%s.log' % other_cmd.replace(' ', ''))
+                switch.cmd('%s %s > %s' % (self.VSCTL, other_cmd, other_dump_name))
             switch.cmd('%s del-br %s' % (self.VSCTL, switch.name))
         self._stop_net()
         self.net = None
