+++ tests/integration/mininet_tests.py
@@ -6578,6 +6578,1494 @@ vlans:
             self.coldstart_conf()
 
 
+class FaucetStringOfDPTest(FaucetTest):
+
+    MAX_HOSTS = 4
+    NUM_HOSTS = 4
+    LINKS_PER_HOST = 1
+    VID = 100
+    CONFIG = None
+    GROUP_TABLE = False
+    dpids = None
+    topo = None
+
+    def non_host_links(self, dpid):
+        return self.topo.dpid_peer_links(dpid)
+
+    @staticmethod
+    def get_config_header(_config_global, _debug_log, _dpid, _hardware):
+        """Don't generate standard config file header."""
+        return ''
+
+    @staticmethod
+    def acls():
+        return {}
+
+    @staticmethod
+    def acl_in_dp():
+        return {}
+
+    def build_net(self, stack=False, n_dps=1,
+                  n_tagged=0, tagged_vid=100,
+                  untagged_hosts=None,
+                  include=None, include_optional=None,
+                  switch_to_switch_links=1, hw_dpid=None,
+                  stack_ring=False, lacp=False, use_external=False,
+                  router=None, dp_options=None):
+        """Set up Mininet and Faucet for the given topology."""
+        if include is None:
+            include = []
+        if include_optional is None:
+            include_optional = []
+        self.dpids = [str(self.rand_dpid()) for _ in range(n_dps)]
+        self.dpids[0] = self.dpid
+        self.topo = mininet_test_topo.FaucetStringOfDPSwitchTopo(
+            self.OVS_TYPE,
+            self.ports_sock,
+            dpids=self.dpids,
+            n_tagged=n_tagged,
+            tagged_vid=tagged_vid,
+            untagged_hosts=untagged_hosts,
+            links_per_host=self.LINKS_PER_HOST,
+            switch_to_switch_links=switch_to_switch_links,
+            test_name=self._test_name(),
+            hw_dpid=hw_dpid, switch_map=self.switch_map,
+            stack_ring=stack_ring,
+            port_order=self.port_order
+        )
+        self.port_maps = {dpid: self.create_port_map(dpid) for dpid in self.dpids}
+        self.port_map = self.port_maps[self.dpid]
+        self.CONFIG = self.get_config(
+            self.dpids,
+            hw_dpid,
+            stack,
+            self.hardware,
+            self.debug_log_path,
+            n_tagged,
+            tagged_vid,
+            untagged_hosts,
+            include,
+            include_optional,
+            self.acls(),
+            self.acl_in_dp(),
+            lacp,
+            use_external,
+            router,
+            dp_options
+        )
+
+    def get_config(self, dpids=None, hw_dpid=None, stack=False, hardware=None, ofchannel_log=None,
+                   n_tagged=0, tagged_vid=0, untagged_hosts=None,
+                   include=None, include_optional=None, acls=None, acl_in_dp=None,
+                   lacp=False, use_external=False, router=None, dp_options=None):
+        """Build a complete Faucet configuration for each datapath, using the given topology."""
+        if dpids is None:
+            dpids = []
+        if include is None:
+            include = []
+        if include_optional is None:
+            include_optional = []
+        if acls is None:
+            acls = {}
+        if acl_in_dp is None:
+            acl_in_dp = {}
+        dpid_names = {}
+        dpname_to_dpkey = {}
+
+        def dp_name(i):
+            return 'faucet-%i' % (i + 1)
+
+        def add_vlans(n_tagged, tagged_vid, untagged_hosts, router):
+            vlans_config = {}
+            if untagged_hosts:
+                for vid in untagged_hosts.keys():
+                    vlans_config[vid] = {
+                        'description': 'untagged',
+                    }
+            if ((n_tagged and not untagged_hosts) or
+                    (n_tagged and untagged_hosts and tagged_vid not in untagged_hosts)):
+                vlans_config[tagged_vid] = {
+                    'description': 'tagged',
+                }
+            if router:
+                for vid in router.keys():
+                    if vid in vlans_config:
+                        if 'faucet_mac' in router[vid]:
+                            vlans_config[vid]['faucet_mac'] = router[vid]['faucet_mac']
+                        if 'faucet_vips' in router[vid]:
+                            vlans_config[vid]['faucet_vips'] = router[vid]['faucet_vips']
+            return vlans_config
+
+        def add_router(router):
+            router_config = {}
+            if router:
+                router_config['router-1'] = {
+                    'vlans': list(router.keys()),
+                }
+            return router_config
+
+        def add_acl_to_port(name, port, interfaces_config):
+            if name in acl_in_dp and port in acl_in_dp[name]:
+                interfaces_config[port]['acl_in'] = acl_in_dp[name][port]
+
+        def add_dp_to_dp_ports(name, dpid, dp_config, interfaces_config, stack,
+                               n_tagged, tagged_vid, untagged_hosts):
+            for link in self.topo.dpid_peer_links(dpid):
+                port, peer_dpid, peer_port = link.port, link.peer_dpid, link.peer_port
+                interfaces_config[port] = {}
+                if stack:
+                    # make this a stacking link.
+                    interfaces_config[port].update(
+                        {
+                            'stack': {
+                                'dp': dpid_names[peer_dpid],
+                                'port': peer_port}
+                        })
+                else:
+                    # not a stack - make this a trunk.
+                    tagged_vlans = []
+                    if n_tagged:
+                        tagged_vlans.append(tagged_vid)
+                    if untagged_hosts:
+                        for vid in untagged_hosts.keys():
+                            if vid not in tagged_vlans:
+                                tagged_vlans.append(vid)
+                    if tagged_vlans:
+                        interfaces_config[port]['tagged_vlans'] = tagged_vlans
+                    if lacp:
+                        interfaces_config[port].update(
+                            {'lacp': 1, 'lacp_active': True})
+                add_acl_to_port(name, port, interfaces_config)
+
+            # TODO: make per test configurable
+            dp_config['lacp_timeout'] = 10
+
+            # TODO: make the stacking root configurable
+            first_dp = dpid == self.dpid
+            if stack and first_dp:
+                dp_config['stack'] = {
+                    'priority': 1
+                }
+
+        def add_dp(name, dpid, hw_dpid, i, stack,
+                   n_tagged, tagged_vid, untagged_hosts,
+                   use_external, dp_options):
+            dp_config = {
+                'dp_id': int(dpid),
+                'hardware': hardware if dpid == hw_dpid else 'Open vSwitch',
+                'ofchannel_log': ofchannel_log + str(i) if ofchannel_log else None,
+                'interfaces': {},
+                'group_table': self.GROUP_TABLE,
+            }
+
+            interfaces_config = {}
+            index = 1
+
+            for n_port in range(n_tagged):
+                port = self.port_maps[dpid]['port_%d' % index]
+                interfaces_config[port] = {
+                    'tagged_vlans': [tagged_vid],
+                    'loop_protect_external': (use_external and n_port != n_tagged - 1),
+                }
+                add_acl_to_port(name, port, interfaces_config)
+                index += 1
+
+            if untagged_hosts:
+                n_port = 0
+                for vid, num_hosts in untagged_hosts.items():
+                    for _ in range(num_hosts):
+                        port = self.port_maps[dpid]['port_%d' % index]
+                        interfaces_config[port] = {
+                            'native_vlan': vid,
+                            'loop_protect_external': (use_external and n_port != num_hosts - 1),
+                        }
+                        add_acl_to_port(name, port, interfaces_config)
+                        index += 1
+                        n_port += 1
+
+            add_dp_to_dp_ports(name, dpid, dp_config, interfaces_config, stack,
+                               n_tagged, tagged_vid, untagged_hosts)
+
+            for portno, config in list(interfaces_config.items()):
+                stack = config.get('stack', None)
+                if stack and 'stack' in interfaces_config[portno]:
+                    peer_portno = stack['port']
+                    interfaces_config[portno]['stack'].update({
+                        'port': 'b%u' % peer_portno})
+
+            dp_config['interfaces'] = interfaces_config
+            if dp_options:
+                for key, value in dp_options.items():
+                    dp_config[key] = value
+            return dp_config
+
+        # Create config
+        config = {'version': 2}
+
+        if include:
+            config['include'] = list(include)
+
+        if include_optional:
+            config['include-optional'] = list(include_optional)
+
+        config['vlans'] = add_vlans(n_tagged, tagged_vid, untagged_hosts, router)
+        if router:
+            config['routers'] = add_router(router)
+        config['acls'] = acls.copy()
+        config['dps'] = {}
+
+        for i, dpid in enumerate(dpids):
+            dpid_names[dpid] = name = dp_name(i)
+            dpname_to_dpkey[name] = dpid
+        self.set_dpid_names(dpid_names)
+
+        for i, dpid in enumerate(dpids):
+            name = dpid_names[dpid]
+            config['dps'][name] = add_dp(
+                name, dpid, hw_dpid, i, stack,
+                n_tagged, tagged_vid, untagged_hosts,
+                use_external, dp_options)
+        config_text = yaml.dump(config, default_flow_style=False)
+        return config_text
+
+    def verify_no_cable_errors(self):
+        i = 0
+        for dpid in self.dpids:
+            i += 1
+            labels = {'dp_id': '0x%x' % int(dpid), 'dp_name': 'faucet-%u' % i}
+            self.assertEqual(
+                0, self.scrape_prometheus_var(
+                    var='stack_cabling_errors_total', labels=labels, default=None))
+            self.assertGreater(
+                self.scrape_prometheus_var(
+                    var='stack_probes_received_total', labels=labels), 0)
+
+    def verify_stack_hosts(self, verify_bridge_local_rule=True, retries=3):
+        lldp_cap_files = []
+        for host in self.hosts_name_ordered():
+            lldp_cap_file = os.path.join(self.tmpdir, '%s-lldp.cap' % host)
+            lldp_cap_files.append(lldp_cap_file)
+            host.cmd(mininet_test_util.timeout_cmd(
+                'tcpdump -U -n -c 1 -i %s -w %s ether proto 0x88CC and not ether src %s &' % (
+                    host.defaultIntf(), host.MAC(), lldp_cap_file), 60))
+        # should not flood LLDP from hosts
+        self.verify_lldp_blocked(self.hosts_name_ordered())
+        # hosts should see no LLDP probes
+        self.verify_empty_caps(lldp_cap_files)
+        if verify_bridge_local_rule:
+            # Verify 802.1x flood block triggered.
+            for dpid in self.dpids:
+                self.wait_nonzero_packet_count_flow(
+                    {'dl_dst': '01:80:c2:00:00:00/ff:ff:ff:ff:ff:f0'},
+                    dpid=dpid, table_id=self._FLOOD_TABLE, ofa_match=False)
+        self.retry_net_ping(retries=retries)
+
+    def stack_port_status(self, dpid, dp_name, port_no):
+        labels = self.port_labels(port_no)
+        labels.update({'dp_id': '0x%x' % int(dpid), 'dp_name': dp_name})
+        return self.scrape_prometheus_var(
+            'port_stack_state', labels=labels,
+            default=None, dpid=False)
+
+    def wait_for_stack_port_status(self, dpid, dp_name, port_no, status, timeout=25):
+        labels = self.port_labels(port_no)
+        labels.update({'dp_id': '0x%x' % int(dpid), 'dp_name': dp_name})
+        if not self.wait_for_prometheus_var(
+                'port_stack_state', status, labels=labels,
+                default=None, dpid=False, timeout=timeout):
+            self.fail('did not get expected dpid %x port %u port_stack_state %u' % (
+                int(dpid), port_no, status))
+
+    def one_stack_port_down(self, dpid, dp_name, port):
+        self.set_port_down(port, dpid, wait=False)
+        self.wait_for_stack_port_status(dpid, dp_name, port, 4)
+
+    def one_stack_port_up(self, dpid, dp_name, port):
+        self.set_port_up(port, dpid, wait=False)
+        self.wait_for_stack_port_status(dpid, dp_name, port, 3)
+
+    def verify_stack_up(self, prop=1.0, timeout=25):
+        for _ in range(timeout):
+            links = 0
+            links_up = 0
+            for i, dpid in enumerate(self.dpids, start=1):
+                dp_name = 'faucet-%u' % i
+                for link in self.non_host_links(dpid):
+                    status = self.stack_port_status(dpid, dp_name, link.port)
+                    links += 1
+                    if status == 3:  # up
+                        links_up += 1
+            prop_up = links_up / links
+            if prop_up >= prop:
+                return
+            time.sleep(1)
+        self.fail('not enough links up: %f / %f' % (links_up, links))
+
+    def verify_one_stack_down(self, stack_offset_port, coldstart=False):
+        self.retry_net_ping()
+        stack_port = self.non_host_links(self.dpid)[stack_offset_port].port
+        remote_stack_port = self.non_host_links(self.dpid)[stack_offset_port].peer_port
+        self.set_port_down(stack_port, wait=False)
+        # self.dpids[1] is the intermediate switch.
+        self.set_port_down(remote_stack_port, self.dpids[1], wait=False)
+        # test case where one link is down when coldstarted.
+        if coldstart:
+            self.coldstart_conf()
+        self.verify_stack_up(prop=0.75)
+        self.verify_stack_hosts(verify_bridge_local_rule=False)
+        # Broadcast works, and first switch doesn't see broadcast packet ins from stack.
+        packet_in_before_broadcast = self.scrape_prometheus_var('of_vlan_packet_ins')
+        self.verify_broadcast()
+        packet_in_after_broadcast = self.scrape_prometheus_var('of_vlan_packet_ins')
+        self.assertEqual(
+            packet_in_before_broadcast,
+            packet_in_after_broadcast)
+        # TODO: re-enable.
+        # self.verify_no_cable_errors()
+
+    def verify_no_arp_storm(self, ping_host, tcpdump_host):
+        num_arp_expected = self.topo.switch_to_switch_links * 2
+        tcpdump_filter = 'arp and ether src %s' % ping_host.MAC()
+        tcpdump_txt = self.tcpdump_helper(
+            tcpdump_host, tcpdump_filter, [
+                lambda: ping_host.cmd('arp -d %s' % tcpdump_host.IP()),
+                lambda: ping_host.cmd(' '.join((self.FPINGS_ARGS_ONE, tcpdump_host.IP())))],
+            packets=(num_arp_expected+1))
+        num_arp_received = len(re.findall(
+            'who-has %s tell %s' % (tcpdump_host.IP(), ping_host.IP()), tcpdump_txt))
+        self.assertTrue(num_arp_received)
+        self.assertLessEqual(num_arp_received, num_arp_expected)
+
+    def verify_stack_has_no_loop(self):
+        for ping_host, tcpdump_host in (
+                (self.hosts_name_ordered()[0], self.hosts_name_ordered()[-1]),
+                (self.hosts_name_ordered()[-1], self.hosts_name_ordered()[0])):
+            self.verify_no_arp_storm(ping_host, tcpdump_host)
+
+    def verify_all_stack_hosts(self):
+        for _ in range(2):
+            self.verify_stack_up()
+            self.verify_no_cable_errors()
+            self.verify_stack_hosts()
+            self.verify_traveling_dhcp_mac()
+            self.verify_unicast_not_looped()
+            self.verify_no_bcast_to_self()
+            self.verify_stack_has_no_loop()
+            self.flap_all_switch_ports()
+
+    def verify_tunnel_established(self, src_host, dst_host, other_host, packets=3):
+        """Verify ICMP packets tunnelled from src to dst."""
+        icmp_match = {'eth_type': IPV4_ETH, 'ip_proto': 1}
+        self.wait_until_matching_flow(icmp_match, table_id=self._PORT_ACL_TABLE, ofa_match=False)
+        tcpdump_text = self.tcpdump_helper(
+            dst_host, 'icmp[icmptype] == 8', [
+                # need to set static ARP as only ICMP is tunnelled.
+                lambda: src_host.cmd('arp -s %s %s' % (other_host.IP(), other_host.MAC())),
+                lambda: src_host.cmd('ping -c%u -t1 %s' % (packets, other_host.IP()))
+            ],
+            packets=1, timeout=(packets + 1),
+        )
+        self.wait_nonzero_packet_count_flow(
+            icmp_match, table_id=self._PORT_ACL_TABLE, ofa_match=False)
+        self.assertTrue(re.search(
+            '%s: ICMP echo request' % other_host.IP(), tcpdump_text
+        ), 'Tunnel was not established')
+
+    def verify_one_broadcast(self, from_host, to_hosts):
+        self.assertGreater(len(to_hosts), 1, 'Testing only one ext host is not useful')
+        received_broadcasts = []
+        for to_host in to_hosts:
+            if self.verify_broadcast(hosts=(from_host, to_host), broadcast_expected=None):
+                received_broadcasts.append(to_host)
+        received_names = {host.name: host for host in received_broadcasts}
+        self.assertEqual(len(received_broadcasts), 1,
+                         'Received not exactly one broadcast from %s: %s' %
+                         (from_host.name, received_names))
+
+    def map_int_ext_hosts(self):
+        conf = self._get_faucet_conf()
+        host_name_map = {host.name: host for host in self.hosts_name_ordered()}
+        int_hosts = set()
+        ext_hosts = set()
+        dp_hosts = {}
+        for dp_name, dp_conf in conf['dps'].items():
+            dpid = int(dp_conf['dp_id'])
+            dp_int_hosts = set()
+            dp_ext_hosts = set()
+            for port, p_conf in dp_conf['interfaces'].items():
+                if 'stack' not in p_conf:
+                    host = host_name_map[self.net.topo.dpid_port_host[dpid][port]]
+                    if p_conf.get('loop_protect_external', False):
+                        dp_ext_hosts.add(host)
+                    else:
+                        dp_int_hosts.add(host)
+            dp_hosts[dp_name] = (dp_int_hosts, dp_ext_hosts)
+            int_hosts.update(dp_int_hosts)
+            ext_hosts.update(dp_ext_hosts)
+        return int_hosts, ext_hosts, dp_hosts
+
+    def verify_protected_connectivity(self):
+        self.verify_stack_up()
+        int_hosts, ext_hosts, dp_hosts = self.map_int_ext_hosts()
+
+        for int_host in int_hosts:
+            # All internal hosts can reach other internal hosts.
+            for other_int_host in int_hosts - {int_host}:
+                self.verify_broadcast(hosts=(int_host, other_int_host), broadcast_expected=True)
+                self.one_ipv4_ping(int_host, other_int_host.IP())
+
+            # All internal hosts can reach exactly one external host.
+            self.verify_one_broadcast(int_host, ext_hosts)
+
+        for ext_host in ext_hosts:
+            # All external hosts cannot flood to each other.
+            for other_ext_host in ext_hosts - {ext_host}:
+                self.verify_broadcast(hosts=(ext_host, other_ext_host), broadcast_expected=False)
+
+            # All external hosts can reach internal hosts.
+            for int_host in int_hosts:
+                self.verify_broadcast(hosts=(ext_host, int_host), broadcast_expected=True)
+                self.one_ipv4_ping(ext_host, int_host.IP())
+
+    def set_externals_state(self, dp_name, externals_up):
+        """Set the port up/down state of all external ports on a switch"""
+        dp_conf = self._get_faucet_conf()['dps'][dp_name]
+        for port_num, port_conf in dp_conf['interfaces'].items():
+            if port_conf.get('loop_protect_external'):
+                if externals_up:
+                    self.set_port_up(port_num, dp_conf.get('dp_id'))
+                else:
+                    self.set_port_down(port_num, dp_conf.get('dp_id'))
+
+    def validate_with_externals_down(self, dp_name):
+        """Check situation when all externals on a given dp are down"""
+        self.set_externals_state(dp_name, False)
+        self.verify_protected_connectivity()
+        self.set_externals_state(dp_name, True)
+
+    def validate_with_externals_down_fails(self, dp_name):
+        """Faucet code is not currently correct, so expect to fail."""
+        # TODO: Fix faucet so the test inversion is no longer required.
+        asserted = False
+        try:
+            self.validate_with_externals_down(dp_name)
+        except AssertionError:
+            asserted = True
+        self.assertTrue(asserted, 'Did not fail as expected for %s' % dp_name)
+
+
+class FaucetSingleUntaggedIPV4RoutingWithStackingTest(FaucetStringOfDPTest):
+    """IPV4 intervlan routing with stacking test"""
+
+    IPV = 4
+    NETPREFIX = 24
+    ETH_TYPE = IPV4_ETH
+    SWITCH_TO_SWITCH_LINKS = 1
+
+    NUM_DPS = 4
+
+    V100 = 100
+    V200 = 200
+    V300 = 300
+
+    V100_NUM_HOSTS = 1
+    V200_NUM_HOSTS = 1
+    V300_NUM_HOSTS = 0
+
+    FAUCET_MAC2 = '0e:00:00:00:00:02'
+
+    @staticmethod
+    def get_dp_options():
+        return {
+            'drop_spoofed_faucet_mac': False,
+            'arp_neighbor_timeout': 2,
+            'max_resolve_backoff_time': 2,
+            'proactive_learn_v4': True
+        }
+
+    def setUp(self):
+        pass
+
+    def set_up(self):
+        super(FaucetSingleUntaggedIPV4RoutingWithStackingTest, self).setUp()
+        router_info = {
+            self.V100: {
+                'faucet_mac': self.FAUCET_MAC,
+                'faucet_vips': [self.get_faucet_vip(1)],
+                'targeted_gw_resolution': False,
+            },
+            self.V200: {
+                'faucet_mac': self.FAUCET_MAC2,
+                'faucet_vips': [self.get_faucet_vip(2)],
+                'targeted_gw_resolution': False,
+            }
+        }
+        untagged_hosts = {self.V100: self.V100_NUM_HOSTS,
+                          self.V200: self.V200_NUM_HOSTS,
+                          self.V300: self.V300_NUM_HOSTS}
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            untagged_hosts=untagged_hosts,
+            switch_to_switch_links=self.SWITCH_TO_SWITCH_LINKS,
+            hw_dpid=self.hw_dpid,
+            router=router_info,
+            dp_options=self.get_dp_options()
+        )
+        self.start_net()
+
+    @staticmethod
+    def get_faucet_mac(vindex):
+        """Get the faucet MAC"""
+        return '0e:00:00:00:00:0%u' % vindex
+
+    def get_faucet_vip(self, vindex):
+        """Get the IPV4 faucet vip"""
+        return '10.%u00.0.254/%u' % (vindex, self.NETPREFIX)
+
+    def get_ip(self, host_n, vindex):
+        """Get the IPV4 host ip"""
+        return '10.%u00.0.%u/%u' % (vindex, host_n, self.NETPREFIX)
+
+    def host_ping(self, src_host, dst_ip):
+        """ping host"""
+        self.one_ipv4_ping(src_host, dst_ip, require_host_learned=False, retries=5)
+
+    def set_host_ip(self, host, host_ip):
+        """Set the host ip"""
+        host.setIP(str(host_ip.ip), prefixLen=self.NETPREFIX)
+
+    def verify_intervlan_routing(self):
+        """Setup host routes and verify intervlan routing is possible"""
+        num_hosts = self.V100_NUM_HOSTS + self.V200_NUM_HOSTS + self.V300_NUM_HOSTS
+        first_faucet_vip = ipaddress.ip_interface(self.get_faucet_vip(1))
+        second_faucet_vip = ipaddress.ip_interface(self.get_faucet_vip(2))
+        v100_hosts = [(self.hosts_name_ordered()[i], ipaddress.ip_interface(
+            self.get_ip(i+1, 1))) for i in range(len(self.hosts_name_ordered())) if (i % num_hosts) == 0]
+        v200_hosts = [(self.hosts_name_ordered()[i], ipaddress.ip_interface(
+            self.get_ip(i+1, 2))) for i in range(len(self.hosts_name_ordered())) if (i % num_hosts) == 1]
+        for host_tuple in v100_hosts:
+            host, host_ip = host_tuple
+            self.set_host_ip(host, host_ip)
+        for host_tuple in v200_hosts:
+            host, host_ip = host_tuple
+            self.set_host_ip(host, host_ip)
+
+        for v100_host_tuple in v100_hosts:
+            v100_host, v100_host_ip = v100_host_tuple
+            for v200_host_tuple in v200_hosts:
+                v200_host, v200_host_ip = v200_host_tuple
+                self.add_host_route(v100_host, v200_host_ip, first_faucet_vip.ip)
+                self.add_host_route(v200_host, v100_host_ip, second_faucet_vip.ip)
+                # TODO: multi DP route resolver needs to flood out stack ports
+                self.host_ping(v100_host, first_faucet_vip.ip)
+                self.host_ping(v200_host, second_faucet_vip.ip)
+                self.host_ping(v100_host, v200_host_ip.ip)
+                self.host_ping(v200_host, v100_host_ip.ip)
+                self.assertEqual(
+                    self._ip_neigh(v100_host, first_faucet_vip.ip, self.IPV), self.FAUCET_MAC)
+                self.assertEqual(
+                    self._ip_neigh(v200_host, second_faucet_vip.ip, self.IPV), self.FAUCET_MAC2)
+        for src_host_tuple in v100_hosts:
+            src_host, _ = src_host_tuple
+            for dst_host_tuple in v100_hosts:
+                _, dst_ip = dst_host_tuple
+                if src_host_tuple == dst_host_tuple:
+                    continue
+                self.host_ping(src_host, dst_ip.ip)
+        for src_host_tuple in v200_hosts:
+            src_host, _ = src_host_tuple
+            for dst_host_tuple in v200_hosts:
+                _, dst_ip = dst_host_tuple
+                if src_host_tuple == dst_host_tuple:
+                    continue
+                self.host_ping(src_host, dst_ip.ip)
+
+    def test_intervlan_routing_2stack(self):
+        """Verify intervlan routing works with 2 DPs in a stack"""
+        self.NUM_DPS = 2
+        self.set_up()
+        self.verify_stack_up()
+        self.verify_intervlan_routing()
+
+    def test_intervlan_routing_3stack(self):
+        """Verify intervlan routing works with 3 DPs in a stack"""
+        self.NUM_DPS = 3
+        self.set_up()
+        self.verify_stack_up()
+        self.verify_intervlan_routing()
+
+    def test_intervlan_routing_4stack(self):
+        """Verify intervlan routing works with 4 DPs in a stack"""
+        self.NUM_DPS = 4
+        self.set_up()
+        self.verify_stack_up()
+        self.verify_intervlan_routing()
+
+    def test_path_no_vlans(self):
+        """Test when a DP in the path of a intervlan route contains no routed VLANs"""
+        self.NUM_DPS = 3
+        self.set_up()
+        first_faucet_vip = ipaddress.ip_interface(self.get_faucet_vip(1))
+        second_faucet_vip = ipaddress.ip_interface(self.get_faucet_vip(2))
+        v100_host = self.hosts_name_ordered()[0]
+        v100_host_ip = ipaddress.ip_interface(self.get_ip(1, 1))
+        v200_host = self.hosts_name_ordered()[5]
+        v200_host_ip = ipaddress.ip_interface(self.get_ip(2, 2))
+        # Remove all hosts on the middle DP by chaning them to hosts on VLAN300
+        #   the middle DP now contains no hosts with VLAN 100 or VLAN 200
+        conf = self._get_faucet_conf()
+        interface_config = conf['dps']['faucet-2']['interfaces']
+        for port_key, port_dict in interface_config.items():
+            if 'stack' in port_dict:
+                continue
+            conf['dps']['faucet-2']['interfaces'][port_key]['native_vlan'] = self.V300
+        self.reload_conf(
+            conf, self.faucet_config_path,
+            restart=True, cold_start=False, change_expected=True)
+        self.verify_stack_up()
+        self.set_host_ip(v100_host, v100_host_ip)
+        self.set_host_ip(v200_host, v200_host_ip)
+        self.add_host_route(v100_host, v200_host_ip, first_faucet_vip.ip)
+        self.add_host_route(v200_host, v100_host_ip, second_faucet_vip.ip)
+        # TODO: multi DP route resolver needs to flood out stack ports
+        self.host_ping(v100_host, first_faucet_vip.ip)
+        self.host_ping(v200_host, second_faucet_vip.ip)
+        self.host_ping(v100_host, v200_host_ip.ip)
+        self.host_ping(v200_host, v100_host_ip.ip)
+        self.assertEqual(
+            self._ip_neigh(v100_host, first_faucet_vip.ip, self.IPV), self.FAUCET_MAC)
+        self.assertEqual(
+            self._ip_neigh(v200_host, second_faucet_vip.ip, self.IPV), self.FAUCET_MAC2)
+
+    def test_dp_one_vlan_from_router(self):
+        """Test when a DP has only one of the routed VLANs"""
+        self.NUM_DPS = 2
+        self.set_up()
+        first_faucet_vip = ipaddress.ip_interface(self.get_faucet_vip(1))
+        second_faucet_vip = ipaddress.ip_interface(self.get_faucet_vip(2))
+        v100_host = self.hosts_name_ordered()[0]
+        v100_host_ip = ipaddress.ip_interface(self.get_ip(1, 1))
+        v200_host = self.hosts_name_ordered()[3]
+        v200_host_ip = ipaddress.ip_interface(self.get_ip(2, 2))
+        # Remove host on VLAN100 by changing it to a host on VLAN300, there is now only
+        #   one host on the DP that is being routed (200)
+        conf = self._get_faucet_conf()
+        interface_config = conf['dps']['faucet-2']['interfaces']
+        for port_key, port_dict in interface_config.items():
+            if 'stack' in port_dict:
+                continue
+            if port_dict['native_vlan'] == self.V100:
+                conf['dps']['faucet-2']['interfaces'][port_key]['native_vlan'] = self.V300
+        self.reload_conf(
+            conf, self.faucet_config_path,
+            restart=True, cold_start=False, change_expected=True)
+        self.verify_stack_up()
+        self.set_host_ip(v100_host, v100_host_ip)
+        self.set_host_ip(v200_host, v200_host_ip)
+        self.add_host_route(v100_host, v200_host_ip, first_faucet_vip.ip)
+        self.add_host_route(v200_host, v100_host_ip, second_faucet_vip.ip)
+        # TODO: multi DP route resolver needs to flood out stack ports
+        self.host_ping(v100_host, first_faucet_vip.ip)
+        self.host_ping(v200_host, second_faucet_vip.ip)
+        self.host_ping(v100_host, v200_host_ip.ip)
+        self.host_ping(v200_host, v100_host_ip.ip)
+        self.assertEqual(
+            self._ip_neigh(v100_host, first_faucet_vip.ip, self.IPV), self.FAUCET_MAC)
+        self.assertEqual(
+            self._ip_neigh(v200_host, second_faucet_vip.ip, self.IPV), self.FAUCET_MAC2)
+
+
+class FaucetSingleUntaggedIPV6RoutingWithStackingTest(FaucetSingleUntaggedIPV4RoutingWithStackingTest):
+    """IPV6 intervlan routing with stacking tests"""
+
+    IPV = 6
+    NETPREFIX = 64
+    ETH_TYPE = IPV6_ETH
+
+    def get_dp_options(self):
+        return {
+            'drop_spoofed_faucet_mac': False,
+            'nd_neighbor_timeout': 2,
+            'max_resolve_backoff_time': 1,
+            'proactive_learn_v6': True
+        }
+
+    def host_ping(self, src_host, dst_ip):
+        self.one_ipv6_ping(src_host, dst_ip, require_host_learned=False)
+
+    def set_host_ip(self, host, host_ip):
+        self.add_host_ipv6_address(host, host_ip)
+
+    def get_faucet_vip(self, vindex):
+        """Get the IPV6 faucet vip"""
+        return 'fc0%u::1:254/112' % vindex
+
+    def get_ip(self, host_n, vindex):
+        """Get the IPV6 host ip"""
+        return 'fc0%u::1:%u/64' % (vindex, host_n)
+
+
+class FaucetStringOfDPUntaggedTest(FaucetStringOfDPTest):
+
+    NUM_DPS = 3
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetStringOfDPUntaggedTest, self).setUp()
+        self.build_net(
+            n_dps=self.NUM_DPS, untagged_hosts={self.VID: self.NUM_HOSTS})
+        self.start_net()
+
+    def test_untagged(self):
+        """All untagged hosts in multi switch topology can reach one another."""
+        self.verify_stack_hosts()
+        self.verify_traveling_dhcp_mac()
+
+
+class FaucetStringOfDPTaggedTest(FaucetStringOfDPTest):
+
+    NUM_DPS = 3
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetStringOfDPTaggedTest, self).setUp()
+        self.build_net(
+            n_dps=self.NUM_DPS, n_tagged=self.NUM_HOSTS, tagged_vid=self.VID)
+        self.start_net()
+
+    def test_tagged(self):
+        """All tagged hosts in multi switch topology can reach one another."""
+        self.verify_stack_hosts(verify_bridge_local_rule=False)
+        self.verify_traveling_dhcp_mac()
+
+
+class FaucetSingleStackStringOfDPTagged0Test(FaucetStringOfDPTest):
+    """Test topology of stacked datapaths with tagged hosts."""
+
+    NUM_DPS = 3
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetSingleStackStringOfDPTagged0Test, self).setUp()
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            n_tagged=self.NUM_HOSTS,
+            tagged_vid=self.VID,
+            switch_to_switch_links=2)
+        self.start_net()
+
+    def test_tagged(self):
+        """All tagged hosts in stack topology can reach each other."""
+        self.verify_stack_up()
+        for coldstart in (False, True):
+            self.verify_one_stack_down(0, coldstart)
+
+
+class FaucetSingleStackStringOfDPTagged1Test(FaucetStringOfDPTest):
+    """Test topology of stacked datapaths with tagged hosts."""
+
+    NUM_DPS = 3
+
+    def setUp(self):  # pylint: disable=invalid-name
+        super(FaucetSingleStackStringOfDPTagged1Test, self).setUp()
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            n_tagged=self.NUM_HOSTS,
+            tagged_vid=self.VID,
+            switch_to_switch_links=2)
+        self.start_net()
+
+    def test_tagged(self):
+        self.verify_stack_up()
+        for coldstart in (False, True):
+            self.verify_one_stack_down(1, coldstart)
+
+
+class FaucetStringOfDPLACPUntaggedTest(FaucetStringOfDPTest):
+    """Test topology of LACP-connected datapaths with untagged hosts."""
+
+    NUM_DPS = 2
+    NUM_HOSTS = 2
+    match_bcast = {'dl_vlan': '100', 'dl_dst': 'ff:ff:ff:ff:ff:ff'}
+    action_str = 'OUTPUT:%u'
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetStringOfDPLACPUntaggedTest, self).setUp()
+        self.build_net(
+            stack=False,
+            n_dps=self.NUM_DPS,
+            untagged_hosts={self.VID: self.NUM_HOSTS},
+            switch_to_switch_links=2,
+            hw_dpid=self.hw_dpid,
+            lacp=True)
+        self.start_net()
+
+    def lacp_ports(self):
+        first_link, second_link = sorted(self.non_host_links(self.dpid))
+        first_lacp_port, second_lacp_port = first_link.port, second_link.port
+        remote_first_lacp_port, remote_second_lacp_port = first_link.peer_port, second_link.peer_port
+        return (first_lacp_port, second_lacp_port,
+                remote_first_lacp_port, remote_second_lacp_port)
+
+    def wait_for_lacp_state(self, port_no, wanted_state, dpid, dp_name, timeout=30):
+        labels = self.port_labels(port_no)
+        labels.update({'dp_id': '0x%x' % int(dpid), 'dp_name': dp_name})
+        if not self.wait_for_prometheus_var(
+                'port_lacp_state', wanted_state,
+                labels=labels, dpid=False, timeout=timeout):
+            self.fail('wanted LACP state for %s to be %u' % (labels, wanted_state))
+
+    def wait_for_lacp_port_init(self, port_no, dpid, dp_name):
+        self.wait_for_lacp_state(port_no, 1, dpid, dp_name)
+
+    def wait_for_lacp_port_up(self, port_no, dpid, dp_name):
+        self.wait_for_lacp_state(port_no, 3, dpid, dp_name)
+
+    def wait_for_lacp_port_noact(self, port_no, dpid, dp_name):
+        self.wait_for_lacp_state(port_no, 5, dpid, dp_name)
+
+    # We sort non_host_links by port because FAUCET sorts its ports
+    # and only floods out of the first active LACP port in that list
+
+    def wait_for_all_lacp_up(self):
+        (first_lacp_port, second_lacp_port, remote_first_lacp_port, _) = self.lacp_ports()
+        self.wait_for_lacp_port_up(first_lacp_port, self.dpid, self.DP_NAME)
+        self.wait_for_lacp_port_up(second_lacp_port, self.dpid, self.DP_NAME)
+        self.wait_until_matching_flow(
+            self.match_bcast, self._FLOOD_TABLE, actions=[self.action_str % first_lacp_port])
+        self.wait_until_matching_flow(
+            self.match_bcast, self._FLOOD_TABLE, actions=[self.action_str % remote_first_lacp_port],
+            dpid=self.dpids[1])
+
+    def test_lacp_port_down(self):
+        """LACP works with any member down."""
+        (first_lacp_port, second_lacp_port,
+         remote_first_lacp_port, remote_second_lacp_port) = self.lacp_ports()
+        local_ports = {first_lacp_port, second_lacp_port}
+        remote_ports = {remote_first_lacp_port, remote_second_lacp_port}
+
+        self.wait_for_all_lacp_up()
+        self.retry_net_ping()
+
+        for local_lacp_port, remote_lacp_port in (
+                (first_lacp_port, remote_first_lacp_port),
+                (second_lacp_port, remote_second_lacp_port)):
+            other_local_lacp_port = list(local_ports - {local_lacp_port})[0]
+            other_remote_lacp_port = list(remote_ports - {remote_lacp_port})[0]
+            self.set_port_down(local_lacp_port, wait=False)
+            self.wait_for_lacp_port_init(
+                local_lacp_port, self.dpid, self.DP_NAME)
+            self.wait_for_lacp_port_init(
+                remote_lacp_port, self.dpids[1], 'faucet-2')
+            self.wait_until_matching_flow(
+                self.match_bcast, self._FLOOD_TABLE, actions=[
+                    self.action_str % other_local_lacp_port])
+            self.wait_until_matching_flow(
+                self.match_bcast, self._FLOOD_TABLE, actions=[
+                    self.action_str % other_remote_lacp_port],
+                dpid=self.dpids[1])
+            self.retry_net_ping()
+            self.set_port_up(local_lacp_port)
+            self.wait_for_all_lacp_up()
+
+    def test_untagged(self):
+        """All untagged hosts in stack topology can reach each other, LAG_CHANGE event emitted."""
+        self._enable_event_log()
+        for _ in range(3):
+            self.wait_for_all_lacp_up()
+            self.verify_stack_hosts()
+            self.flap_all_switch_ports()
+        # Check for presence of LAG_CHANGE event in event socket log
+        self.wait_until_matching_lines_from_file(r'.+LAG_CHANGE.+', self.event_log)
+
+    def test_dyn_fail(self):
+        """Test lacp fail on reload with dynamic lacp status."""
+
+        conf = self._get_faucet_conf()
+        (src_port, dst_port, fail_port, _) = self.lacp_ports()
+
+        self.wait_for_lacp_port_up(src_port, self.dpids[0], 'faucet-1')
+        self.wait_for_lacp_port_up(dst_port, self.dpids[0], 'faucet-1')
+
+        interfaces_conf = conf['dps']['faucet-2']['interfaces']
+        interfaces_conf[fail_port]['lacp'] = 0
+        interfaces_conf[fail_port]['lacp_active'] = False
+        self.reload_conf(conf, self.faucet_config_path, restart=True,
+                         cold_start=False, change_expected=False)
+
+        self.wait_for_lacp_port_init(src_port, self.dpids[0], 'faucet-1')
+        self.wait_for_lacp_port_up(dst_port, self.dpids[0], 'faucet-1')
+
+    def test_passthrough(self):
+        """Test lacp passthrough on port fail."""
+
+        conf = self._get_faucet_conf()
+        (src_port, dst_port, fail_port, end_port) = self.lacp_ports()
+
+        interfaces_conf = conf['dps']['faucet-1']['interfaces']
+        interfaces_conf[dst_port]['lacp_passthrough'] = [src_port]
+        interfaces_conf[dst_port]['loop_protect_external'] = True
+        interfaces_conf[dst_port]['lacp'] = 2
+        interfaces_conf[src_port]['loop_protect_external'] = True
+        interfaces_conf = conf['dps']['faucet-2']['interfaces']
+        interfaces_conf[fail_port]['loop_protect_external'] = True
+        interfaces_conf[end_port]['loop_protect_external'] = True
+        interfaces_conf[end_port]['lacp'] = 2
+
+        self.reload_conf(conf, self.faucet_config_path, restart=True,
+                         cold_start=False, change_expected=False)
+
+        self.wait_for_all_lacp_up()
+        self.verify_stack_hosts()
+
+        interfaces_conf[fail_port]['lacp'] = 0
+        interfaces_conf[fail_port]['lacp_active'] = False
+        self.reload_conf(conf, self.faucet_config_path, restart=True,
+                         cold_start=False, change_expected=False)
+
+        self.wait_for_lacp_port_init(src_port, self.dpids[0], 'faucet-1')
+        self.wait_for_lacp_port_up(dst_port, self.dpids[0], 'faucet-1')
+        self.wait_for_lacp_port_init(end_port, self.dpids[1], 'faucet-2')
+
+
+class FaucetStackStringOfDPUntaggedTest(FaucetStringOfDPTest):
+    """Test topology of stacked datapaths with untagged hosts."""
+
+    NUM_DPS = 2
+    NUM_HOSTS = 2
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetStackStringOfDPUntaggedTest, self).setUp()
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            untagged_hosts={self.VID: self.NUM_HOSTS},
+            switch_to_switch_links=2,
+            hw_dpid=self.hw_dpid)
+        self.start_net()
+
+    def test_untagged(self):
+        """All untagged hosts in stack topology can reach each other."""
+        self.verify_stack_hosts()
+
+
+class FaucetSingleStackStringOfDPExtLoopProtUntaggedTest(FaucetStringOfDPTest):
+    """Test topology of stacked datapaths with untagged hosts."""
+
+    NUM_DPS = 2
+    NUM_HOSTS = 3
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetSingleStackStringOfDPExtLoopProtUntaggedTest, self).setUp()
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            untagged_hosts={self.VID: self.NUM_HOSTS},
+            switch_to_switch_links=2,
+            hw_dpid=self.hw_dpid,
+            use_external=True)
+        self.start_net()
+
+    def test_untagged(self):
+        """Host can reach each other, unless both marked loop_protect_external"""
+        for host in self.hosts_name_ordered():
+            self.require_host_learned(host)
+
+        # Part 1: Make sure things are connected properly.
+        self.verify_protected_connectivity()  # Before reload
+
+        # Part 2: Test the code on pipeline reconfiguration path.
+        conf = self._get_faucet_conf()
+        loop_interface = None
+        for interface, interface_conf in conf['dps']['faucet-2']['interfaces'].items():
+            if 'stack' in interface_conf:
+                continue
+            if not interface_conf.get('loop_protect_external', False):
+                loop_interface = interface
+                break
+
+        self._mark_external(loop_interface, True)
+        self._mark_external(loop_interface, False)
+
+        # Part 3: Make sure things are the same after reload.
+        self.verify_protected_connectivity()  # After reload
+
+    def _mark_external(self, loop_interface, protect_external):
+        conf = self._get_faucet_conf()
+        conf['dps']['faucet-2']['interfaces'][loop_interface]['loop_protect_external'] = protect_external
+        self.reload_conf(
+            conf, self.faucet_config_path,
+            restart=True, cold_start=False, change_expected=True)
+
+    def test_missing_ext(self):
+        """Test stacked dp with all external ports down on a switch"""
+
+        self.validate_with_externals_down_fails('faucet-1')
+        self.validate_with_externals_down_fails('faucet-2')
+
+
+class FaucetSingleStackStringOf3DPExtLoopProtUntaggedTest(FaucetStringOfDPTest):
+    """Test topology of stacked datapaths with untagged hosts."""
+
+    NUM_DPS = 3
+    NUM_HOSTS = 3
+
+    def setUp(self):  # pylint: disable=invalid-name
+        super(FaucetSingleStackStringOf3DPExtLoopProtUntaggedTest, self).setUp()
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            untagged_hosts={self.VID: self.NUM_HOSTS},
+            switch_to_switch_links=2,
+            hw_dpid=self.hw_dpid,
+            use_external=True)
+        self.start_net()
+
+    def test_untagged(self):
+        self.verify_stack_up()
+        int_hosts, ext_hosts, dp_hosts = self.map_int_ext_hosts()
+        _, root_ext_hosts = dp_hosts[self.DP_NAME]
+
+        for int_host in int_hosts:
+            # All internal hosts can reach other internal hosts.
+            for other_int_host in int_hosts - {int_host}:
+                self.verify_broadcast(
+                    hosts=(int_host, other_int_host), broadcast_expected=True)
+                self.verify_unicast(
+                    hosts=(int_host, other_int_host), unicast_expected=True)
+
+            # All internal hosts should reach exactly one external host.
+            self.verify_one_broadcast(int_host, ext_hosts)
+
+        for ext_host in ext_hosts:
+            # All external hosts cannot flood to each other
+            for other_ext_host in ext_hosts - {ext_host}:
+                self.verify_broadcast(
+                    hosts=(ext_host, other_ext_host), broadcast_expected=False)
+
+        remote_ext_hosts = ext_hosts - set(root_ext_hosts)
+        # int host should never be broadcast to an ext host that is not on the root.
+        for local_int_hosts, _ in dp_hosts.values():
+            for local_int_host in local_int_hosts:
+                for remote_ext_host in remote_ext_hosts:
+                    self.verify_broadcast(
+                        hosts=(local_int_host, remote_ext_host), broadcast_expected=False)
+
+
+class FaucetGroupStackStringOfDPUntaggedTest(FaucetStackStringOfDPUntaggedTest):
+    """Test topology of stacked datapaths with untagged hosts."""
+
+    GROUP_TABLE = True
+
+
+class FaucetStackRingOfDPTest(FaucetStringOfDPTest):
+
+    NUM_DPS = 3
+    SOFTWARE_ONLY = True
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetStackRingOfDPTest, self).setUp()
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            untagged_hosts={self.VID: self.NUM_HOSTS},
+            switch_to_switch_links=1,
+            stack_ring=True)
+        self.start_net()
+
+    def test_untagged(self):
+        """Stack loop prevention works and hosts can ping each other, STACK_TOPO_CHANGE event emitted."""
+        self._enable_event_log()
+        self.verify_stack_up()
+        self.verify_stack_has_no_loop()
+        self.retry_net_ping()
+        self.verify_traveling_dhcp_mac()
+        # Move through each DP breaking either side of the ring
+        for dpid_i in range(self.NUM_DPS):
+            dpid = self.dpids[dpid_i]
+            dp_name = 'faucet-%u' % (dpid_i + 1)
+            for link in self.non_host_links(dpid):
+                port = link.port
+                self.one_stack_port_down(dpid, dp_name, port)
+                self.retry_net_ping()
+                self.one_stack_port_up(dpid, dp_name, port)
+        # Check for presence of STACK_TOPO_CHANGE event in event socket log
+        self.wait_until_matching_lines_from_file(r'.+STACK_TOPO_CHANGE.+', self.event_log)
+
+
+class FaucetSingleStack4RingOfDPTest(FaucetStackRingOfDPTest):
+
+    NUM_DPS = 4
+
+
+class FaucetSingleStackAclControlTest(FaucetStringOfDPTest):
+    """Test ACL control of stacked datapaths with untagged hosts."""
+
+    NUM_DPS = 3
+    NUM_HOSTS = 3
+
+    def acls(self):
+        map1, map2, map3 = [self.port_maps[dpid] for dpid in self.dpids]
+        return {
+            1: [
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'nw_dst': '10.0.0.2',
+                    'actions': {
+                        'output': {
+                            'port': map1['port_2']
+                        }
+                    },
+                }},
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'dl_dst': 'ff:ff:ff:ff:ff:ff',
+                    'actions': {
+                        'output': {
+                            'ports': [
+                                map1['port_2'],
+                                map1['port_4']]
+                        }
+                    },
+                }},
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'actions': {
+                        'output': {
+                            'port': map1['port_4']
+                        }
+                    },
+                }},
+                {'rule': {
+                    'actions': {
+                        'allow': 1,
+                    },
+                }},
+            ],
+            2: [
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'actions': {
+                        'output': {
+                            'port': map2['port_5']
+                        }
+                    },
+                }},
+                {'rule': {
+                    'actions': {
+                        'allow': 1,
+                    },
+                }},
+            ],
+            3: [
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'nw_dst': '10.0.0.7',
+                    'actions': {
+                        'output': {
+                            'port': map3['port_1']
+                        }
+                    },
+                }},
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'dl_dst': 'ff:ff:ff:ff:ff:ff',
+                    'actions': {
+                        'output': {
+                            'ports': [map3['port_1']]
+                        }
+                    },
+                }},
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'actions': {
+                        'allow': 0,
+                    },
+                }},
+                {'rule': {
+                    'actions': {
+                        'allow': 1,
+                    },
+                }},
+            ],
+        }
+
+    # DP-to-acl_in port mapping.
+    def acl_in_dp(self):
+        map1, map2, map3 = [self.port_maps[dpid] for dpid in self.dpids]
+        return {
+            'faucet-1': {
+                # Port 1, acl_in = 1
+                map1['port_1']: 1,
+            },
+            'faucet-2': {
+                # Port 4, acl_in = 2
+                map2['port_4']: 2,
+            },
+            'faucet-3': {
+                # Port 4, acl_in = 3
+                map3['port_4']: 3,
+            },
+        }
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetSingleStackAclControlTest, self).setUp()
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            untagged_hosts={self.VID: self.NUM_HOSTS},
+            )
+        self.start_net()
+
+    def test_unicast(self):
+        """Hosts in stack topology can appropriately reach each other over unicast."""
+        hosts = self.hosts_name_ordered()
+        self.verify_stack_up()
+        self.verify_tp_dst_notblocked(5000, hosts[0], hosts[1], table_id=None)
+        self.verify_tp_dst_blocked(5000, hosts[0], hosts[3], table_id=None)
+        self.verify_tp_dst_notblocked(5000, hosts[0], hosts[6], table_id=None)
+        self.verify_tp_dst_blocked(5000, hosts[0], hosts[7], table_id=None)
+        self.verify_no_cable_errors()
+
+    def test_broadcast(self):
+        """Hosts in stack topology can appropriately reach each other over broadcast."""
+        hosts = self.hosts_name_ordered()
+        self.verify_stack_up()
+        self.verify_bcast_dst_notblocked(5000, hosts[0], hosts[1])
+        self.verify_bcast_dst_blocked(5000, hosts[0], hosts[3])
+        self.verify_bcast_dst_notblocked(5000, hosts[0], hosts[6])
+        self.verify_bcast_dst_blocked(5000, hosts[0], hosts[7])
+        self.verify_no_cable_errors()
+
+
+class FaucetStringOfDPACLOverrideTest(FaucetStringOfDPTest):
+
+    NUM_DPS = 1
+    NUM_HOSTS = 2
+
+    # ACL rules which will get overridden.
+    def acls(self):
+        return {
+            1: [
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'ip_proto': 6,
+                    'tcp_dst': 5001,
+                    'actions': {
+                        'allow': 1,
+                    },
+                }},
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'ip_proto': 6,
+                    'tcp_dst': 5002,
+                    'actions': {
+                        'allow': 0,
+                    },
+                }},
+                {'rule': {
+                    'actions': {
+                        'allow': 1,
+                    },
+                }},
+            ],
+        }
+
+    # ACL rules which get put into an include-optional
+    # file, then reloaded into FAUCET.
+    @staticmethod
+    def acls_override():
+        return {
+            1: [
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'ip_proto': 6,
+                    'tcp_dst': 5001,
+                    'actions': {
+                        'allow': 0,
+                    },
+                }},
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'ip_proto': 6,
+                    'tcp_dst': 5002,
+                    'actions': {
+                        'allow': 1,
+                    },
+                }},
+                {'rule': {
+                    'actions': {
+                        'allow': 1,
+                    },
+                }},
+            ],
+        }
+
+    # DP-to-acl_in port mapping.
+    def acl_in_dp(self):
+        port_1 = self.port_map['port_1']
+        return {
+            'faucet-1': {
+                # First port, acl_in = 1
+                port_1: 1,
+            },
+        }
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetStringOfDPACLOverrideTest, self).setUp()
+        self.acls_config = os.path.join(self.tmpdir, 'acls.yaml')
+        missing_config = os.path.join(self.tmpdir, 'missing_config.yaml')
+        self.build_net(
+            n_dps=self.NUM_DPS,
+            untagged_hosts={self.VID: self.NUM_HOSTS},
+            include_optional=[self.acls_config, missing_config],
+        )
+        self.start_net()
+
+    def test_port5001_blocked(self):
+        """Test that TCP port 5001 is blocked."""
+        self.ping_all_when_learned()
+        first_host, second_host = self.hosts_name_ordered()[0:2]
+        self.verify_tp_dst_notblocked(5001, first_host, second_host)
+        with open(self.acls_config, 'w') as config_file:
+            config_file.write(self.get_config(acls=self.acls_override()))
+        self.verify_faucet_reconf(cold_start=False, change_expected=True)
+        self.verify_tp_dst_blocked(5001, first_host, second_host)
+        self.verify_no_cable_errors()
+
+    def test_port5002_notblocked(self):
+        """Test that TCP port 5002 is not blocked."""
+        self.ping_all_when_learned()
+        first_host, second_host = self.hosts_name_ordered()[0:2]
+        self.verify_tp_dst_blocked(5002, first_host, second_host)
+        with open(self.acls_config, 'w') as config_file:
+            config_file.write(self.get_config(acls=self.acls_override()))
+        self.verify_faucet_reconf(cold_start=False, change_expected=True)
+        self.verify_tp_dst_notblocked(5002, first_host, second_host)
+        self.verify_no_cable_errors()
+
+
+class FaucetTunnelSameDpTest(FaucetStringOfDPTest):
+
+    NUM_DPS = 2
+    NUM_HOSTS = 2
+    SWITCH_TO_SWITCH_LINKS = 2
+    VID = 100
+
+    def acls(self):
+        return {
+            1: [
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'ip_proto': 1,
+                    'actions': {
+                        'allow': 0,
+                        'output': {
+                            'tunnel': {
+                                'type': 'vlan',
+                                'tunnel_id': 200,
+                                'dp': 'faucet-1',
+                                'port': 'b%(port_2)d'}
+                        }
+                    }
+                }}
+            ]
+        }
+
+    # DP-to-acl_in port mapping.
+    def acl_in_dp(self):
+        port_1 = self.port_map['port_1']
+        return {
+            'faucet-1': {
+                # First port 1, acl_in = 1
+                port_1: 1,
+            }
+        }
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetTunnelSameDpTest, self).setUp()
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            untagged_hosts={self.VID: self.NUM_HOSTS},
+            switch_to_switch_links=self.SWITCH_TO_SWITCH_LINKS,
+            hw_dpid=self.hw_dpid,
+        )
+        self.start_net()
+
+    def test_tunnel_established(self):
+        """Test a tunnel path can be created."""
+        self.verify_stack_up()
+        src_host, dst_host, other_host = self.hosts_name_ordered()[:3]
+        self.verify_tunnel_established(src_host, dst_host, other_host)
+
+
+class FaucetTunnelTest(FaucetStringOfDPTest):
+
+    NUM_DPS = 2
+    NUM_HOSTS = 2
+    SWITCH_TO_SWITCH_LINKS = 2
+    VID = 100
+
+    def acls(self):
+        dpid2 = self.dpids[1]
+        port2_1 = self.port_maps[dpid2]['port_1']
+        return {
+            1: [
+                {'rule': {
+                    'dl_type': IPV4_ETH,
+                    'ip_proto': 1,
+                    'actions': {
+                        'allow': 0,
+                        'output': {
+                            'tunnel': {
+                                'type': 'vlan',
+                                'tunnel_id': 200,
+                                'dp': 'faucet-2',
+                                'port': port2_1}
+                        }
+                    }
+                }}
+            ]
+        }
+
+    # DP-to-acl_in port mapping.
+    def acl_in_dp(self,):
+        port_1 = self.port_map['port_1']
+        return {
+            'faucet-1': {
+                # First port 1, acl_in = 1
+                port_1: 1,
+            }
+        }
+
+    def setUp(self): # pylint: disable=invalid-name
+        super(FaucetTunnelTest, self).setUp()
+        self.build_net(
+            stack=True,
+            n_dps=self.NUM_DPS,
+            untagged_hosts={self.VID: self.NUM_HOSTS},
+            switch_to_switch_links=self.SWITCH_TO_SWITCH_LINKS,
+            hw_dpid=self.hw_dpid,
+        )
+        self.start_net()
+
+    def test_tunnel_established(self):
+        """Test a tunnel path can be created."""
+        self.verify_stack_up()
+        src_host, other_host, dst_host = self.hosts_name_ordered()[:3]
+        self.verify_tunnel_established(src_host, dst_host, other_host)
+
+    def test_tunnel_path_rerouted(self):
+        """Test a tunnel path is rerouted when a stack is down."""
+        self.verify_stack_up()
+        first_stack_port = self.non_host_links(self.dpid)[0].port
+        self.one_stack_port_down(self.dpid, self.DP_NAME, first_stack_port)
+        src_host, other_host, dst_host = self.hosts_name_ordered()[:3]
+        self.verify_tunnel_established(src_host, dst_host, other_host, packets=10)
+        self.set_port_up(first_stack_port, self.dpid)
+
+
 class FaucetGroupTableTest(FaucetUntaggedTest):
 
     CONFIG = """
