+++ travis/runtests.sh
@@ -45,10 +45,8 @@ if [ "${MATRIX_SHARD}" == "unittest" ] ; then
   exit 0
 elif [ "${MATRIX_SHARD}" == "sanity" ] ; then
   FAUCET_TESTS="-u FaucetSanityTest"
-elif [ "${MATRIX_SHARD}" == "fault-tolerance" ] ; then
-  FAUCET_TESTS="-t"
 else
-  ALLTESTFILES="tests/integration/mininet_tests.py clib/clib_mininet_tests.py tests/integration/mininet_multidp_tests.py"
+  ALLTESTFILES="tests/integration/mininet_tests.py clib/clib_mininet_tests.py"
   ALLTESTS=`grep -E -o "^class (Faucet[a-zA-Z0-9]+Test)" ${ALLTESTFILES}|cut -f2 -d" "|sort`
   declare -A sharded
